# On mutations tab don't wait for gene panel data