@onursumer 2021-05-17
Thanks @inodb! 

A minor thing. Right now we don't see any loading indicator for the `Somatic Mutation Frequency`. It just pops into existence after the loading completes. Should we do something like this instead?

![mutation_rate_spinner](https://user-images.githubusercontent.com/3604198/118557631-763f3100-b733-11eb-9b90-7f61ad83aa9a.png)