@inodb 2021-05-17
@onursumer yeah good point, that would be nice - filed an issue  https://github.com/cBioPortal/cbioportal/issues/8606