@inodb 2021-05-18
Also found a bug related to selecting germline/mutation rate: https://github.com/cBioPortal/cbioportal/issues/8607